+++ services/api-server/src/simcore_service_api_server/__main__.py
@@ -8,14 +8,15 @@ from pathlib import Path
 
 import uvicorn
 from fastapi import FastAPI
-from simcore_service_api_server.core.settings import AppSettings, BootModeEnum
+
 from simcore_service_api_server.core.application import init_app
+from simcore_service_api_server.core.settings import AppSettings, BootModeEnum
 
 current_dir = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
 
 
 # SINGLETON FastAPI app
-the_app: FastAPI =  init_app()
+the_app: FastAPI = init_app()
 
 
 def main():
@@ -28,6 +29,7 @@ def main():
         reload_dir=current_dir,
         log_level=cfg.log_level_name.lower(),
     )
+
 
 if __name__ == "__main__":
     main()
